refactor: rename app directory to store
