+++ rs-react-app/components/CardList.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Card } from './Card';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../app/store';
+import { RootState } from '../store/store';
 
 export const CardList: FC = () => {
   const { results } = useSelector((state: RootState) => state.results);
