+++ rs-react-app/components/Flayout.test.tsx
@@ -5,7 +5,7 @@ import { Flayout } from './Flayout';
 import { vi, describe, test, expect, beforeEach } from 'vitest';
 import { ThemeContext } from '../context/ThemeContext';
 import '@testing-library/jest-dom';
-import { AppDispatch, RootState } from '../app/store';
+import { AppDispatch, RootState } from '../store/store';
 import { deleteChecks, fetchItemDetails } from '../features/check';
 
 const mockStore = configureStore<RootState, AppDispatch>([]);
