+++ rs-react-app/components/customRender.tsx
@@ -1,5 +1,5 @@
 import { ThemeProvider } from '../context/ThemeContext';
-import { store } from '../app/store';
+import { store } from '../store/store';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 import { ReactNode } from 'react';
