+++ rs-react-app/next-env.d.ts
@@ -1,15 +1,5 @@
 /// <reference types="next" />
 /// <reference types="next/image-types/global" />
-/// <reference types="next/navigation-types/compat/navigation" />
 
 // NOTE: This file should not be edited
-// see https://nextjs.org/docs/app/api-reference/config/typescript for more information.
-interface Window {
-  showSaveFilePicker(options?: {
-    suggestedName?: string;
-    types?: Array<{
-      description?: string;
-      accept: Record<string, string[]>;
-    }>;
-  }): Promise<FileSystemFileHandle>;
-}
+// see https://nextjs.org/docs/pages/api-reference/config/typescript for more information.
