+++ rs-react-app/pages/details/[id].tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
 
 import { useTheme } from '../../context/ThemeContext';
-import '../../styles/App.css';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
 import { Header } from '../../components/Header';
 import { Main } from '../../components/Main';
